testing/web-platform/tests/html/semantics/forms/the-datalist-element/datalistelement.html

<!DOCTYPE html>
<html><head>
    <title>datalist element</title>
    <meta content="text/html; charset=UTF-8" http-equiv="content-type">
    <meta content="datalist element" name="description">
    <link href="http://www.w3.org/TR/html5/forms.html#the-datalist-element" rel="help">
  </head>
  <body>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <div id="log"></div>

    <datalist id="datalist1" style="display: none">
      <option>1</option>
      <option>2</option>
      <option>3</option>
      <option>4</option>
    </datalist>

    <script type="text/javascript">
      test(function() {assert_true(document.getElementsByTagName("datalist")[0] instanceof HTMLDataListElement)}, "HTMLDataListElement support");
      test(function() {assert_equals(document.getElementsByTagName("datalist")[0].options.item(0), document.getElementsByTagName("option")[0])}, "HTMLDataListElement options item collection support");
    </script>

</body></html>
